Exercícios do Curso em Vídeo

+++ pythonExercicios/ex033.py
@@ -1,6 +1,6 @@
-num1 = int(input('Digite o primeiro número: '))
-num2 = int(input('Digite o segundo número: '))
-num3 = int(input('Digite o terceiro número: '))
+num1 = float(input('Digite o primeiro número: '))
+num2 = float(input('Digite o segundo número: '))
+num3 = float(input('Digite o terceiro número: '))
 if num1 == num2 == num3:
     print('Os três números são IGUAIS!')
 elif num1 >= num2 and num1 >= num3 and num2 <= num3:
